requirements.txt
  torch==2.4.1
  torchaudio==2.4.1
  torchvision==0.19.1
  torchtext
  transformers<4.47.0
  ipython
  einops
  biotite==0.41.2
  msgpack-numpy
  biopython
  scikit-learn
  brotli
  attrs
  pandas
  cloudpathlib
  tenacity
  wandb
  torcheval
  mdtraj
  MDAnalysis
